requirements.txt
requests
pyyaml
tk
customtkinter